+++ packages/react/test/specs/components/Animation/Animation-test.tsx
@@ -3,5 +3,5 @@ import { isConformant } from 'test/specs/commonTests'
 import Animation from 'src/components/Animation/Animation'
 
 describe('Animation', () => {
-  isConformant(Animation)
+  isConformant(Animation, { hasAccessibilityProp: false })
 })